Added camera movement to the portal
- Portal camera follows the player's camera.

+++ ProgGrafica_2024/lib/Render.h
@@ -42,7 +42,7 @@ class Render
 
 		void setupObject(Object* object);
 
-		void move(float deltaTime);
+		void move(double deltaTime);
 
 		void drawObjects();
